Refactor multipart upload abstraction to isolate interaction with streams
